v1.6.0

+++ lib/helpers/attachment_downloader.dart
@@ -1,4 +1,10 @@
-import 'dart:io';
+import 'dart:convert';
+
+import 'package:bluebubbles/helpers/utils.dart';
+import 'package:bluebubbles/managers/current_chat.dart';
+import 'package:bluebubbles/repository/models/platform_file.dart';
+import 'package:flutter/foundation.dart';
+import 'package:universal_io/io.dart';
 
 import 'package:bluebubbles/helpers/attachment_helper.dart';
 import 'package:bluebubbles/helpers/logger.dart';
@@ -36,7 +42,7 @@ class AttachmentDownloadController extends GetxController {
   final Function? onComplete;
   final Function? onError;
   final RxnNum progress = RxnNum();
-  final Rxn<File> file = Rxn<File>();
+  final Rxn<PlatformFile> file = Rxn<PlatformFile>();
   final RxBool error = RxBool(false);
   int chunkSize = 500;
   Stopwatch stopwatch = Stopwatch();
@@ -64,19 +70,47 @@ class AttachmentDownloadController extends GetxController {
     getChunkRecursive(attachment.guid!, 0, numOfChunks, []);
   }
 
-  void getChunkRecursive(String guid, int index, int total, List<int> currentBytes) {
+  Future<void> getChunkRecursive(String guid, int index, int total, List<int> currentBytes) async {
     // if (index <= total) {
     Map<String, dynamic> params = new Map();
     params["identifier"] = guid;
     params["start"] = index * chunkSize;
     params["chunkSize"] = chunkSize;
     params["compress"] = false;
+    if (kIsWeb) {
+      var response = await api.downloadAttachment(attachment.guid!);
+      attachment.webUrl = response.requestOptions.path;
+      Logger.info("Finished fetching attachment");
+      stopwatch.stop();
+      Logger.info("Attachment downloaded in ${stopwatch.elapsedMilliseconds} ms");
+
+      if (CurrentChat.activeChat?.chatAttachments.firstWhereOrNull((e) => e.guid == attachment.guid) ==
+          null) {
+        CurrentChat.activeChat?.chatAttachments.add(attachment);
+      }
+
+      // Finish the downloader
+      Get.find<AttachmentDownloadService>().removeFromQueue(this);
+      if (onComplete != null) onComplete!();
+      attachment.bytes = response.data;
+      // Add attachment to sink based on if we got data
+
+      file.value = PlatformFile(
+        name: attachment.transferName!,
+        path: kIsWeb ? null : attachment.getPath(),
+        size: total,
+        bytes: response.data,
+      );
+      return;
+    }
     SocketManager().sendMessage("get-attachment-chunk", params, (attachmentResponse) async {
       if (attachmentResponse['status'] != 200 ||
           (attachmentResponse.containsKey("error") && attachmentResponse["error"] != null)) {
-        File file = new File(attachment.getPath());
-        if (await file.exists()) {
-          await file.delete();
+        if (!kIsWeb) {
+          File file = new File(attachment.getPath());
+          if (await file.exists()) {
+            await file.delete();
+          }
         }
 
         if (onError != null) onError!.call();
@@ -88,7 +122,7 @@ class AttachmentDownloadController extends GetxController {
 
       int? numBytes = attachmentResponse["byteLength"];
 
-      if (numBytes == chunkSize) {
+      if (numBytes == chunkSize && (progress.value ?? 0) < 1) {
         // Calculate some stats
         double progress = ((index + 1) / total).clamp(0, 1).toDouble();
         Logger.info("Progress: ${(progress * 100).round()}% of the attachment");
@@ -105,22 +139,37 @@ class AttachmentDownloadController extends GetxController {
 
         try {
           // Compress the attachment
-          await AttachmentHelper.compressAttachment(attachment, attachment.getPath());
-          await attachment.update();
+          if (!kIsWeb) {
+            await AttachmentHelper.compressAttachment(attachment, attachment.getPath());
+            await attachment.update();
+          } else if (CurrentChat.activeChat?.chatAttachments.firstWhereOrNull((e) => e.guid == attachment.guid) ==
+              null) {
+            CurrentChat.activeChat?.chatAttachments.add(attachment);
+          }
         } catch (ex) {
           // So what if it crashes here.... I don't care...
         }
 
-        File downloadedFile = new File(attachment.getPath());
-
         // Finish the downloader
         Get.find<AttachmentDownloadService>().removeFromQueue(this);
         if (onComplete != null) onComplete!();
-
+        attachment.bytes = base64.decode(attachmentResponse['data']);
         // Add attachment to sink based on if we got data
-        file.value = downloadedFile;
+
+        file.value = PlatformFile(
+          name: attachment.transferName!,
+          path: kIsWeb ? null : attachment.getPath(),
+          size: total,
+          bytes: attachment.bytes,
+        );
+        if (kIsDesktop) {
+          if (attachment.bytes != null) {
+            File _file = await File(attachment.getPath()).create(recursive: true);
+            _file.writeAsBytesSync(attachment.bytes!.toList());
+          }
+        }
       }
-    }, reason: "Attachment downloader " + attachment.guid!, path: attachment.getPath());
+    }, reason: "Attachment downloader " + attachment.guid!, path: kIsWeb ? "" : attachment.getPath());
   }
 
   void setProgress(double value) {
